# GODRIVER-2914 bsoncodec/bsonrw: eliminate encoding allocations