@charlievieth 2023-09-05
SGTM

On Tue, Sep 5, 2023 at 3:29 PM Matt Dale ***@***.***> wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In bson/bsoncodec/struct_codec.go
> <https://github.com/mongodb/mongo-go-driver/pull/1323#discussion_r1316311875>
> :
>
> > @@ -393,7 +393,7 @@ func (sc *StructCodec) DecodeValue(dc DecodeContext, vr bsonrw.ValueReader, val
>
>  func isZero(v reflect.Value, omitZeroStruct bool) bool {
>  	kind := v.Kind()
> -	if (kind != reflect.Pointer || !v.IsNil()) && v.Type().Implements(tZeroer) {
> +	if (kind != reflect.Ptr || !v.IsNil()) && v.Type().Implements(tZeroer) {
>
> You're correct. However, the Go driver currently supports a min Go version
> of 1.13, which doesn't have the new Pointer symbol, so we have to use the
> old Ptr symbol.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/mongodb/mongo-go-driver/pull/1323#discussion_r1316311875>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABFGHBWILJZOWNFJ4TDUTETXY54QLANCNFSM6AAAAAA2LAQCAY>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>